+++ tinper-bee/bee-datepicker/src/DatePicker.js
@@ -141,7 +141,7 @@ class DatePicker extends Component {
   handleCalendarChange = (value) => {
       const props = this.props;
       this.setState({ value: value,inputValue:(value && this.getValue(value)) || '' });
-      this.fireChange(value, (value && this.getValue(value)) || '');
+      // this.fireChange(value, (value && this.getValue(value)) || '');
   }
   handleChange = value => {
     const props = this.props;